+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## master
+
+* `derivatives` – Add `:io` and `:attacher` values to instrumentation event payload (@janko)
+
 ## 3.1.0 (2019-11-15) [[release notes]](https://shrinerb.com/docs/release_notes/3.1.0)
 
 * `default_storage` – Coerce storage key to symbol in `Attacher#cache_key` & `Attacher#store_key` (@janko)